requirements.txt
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b
requests
openai
python-dotenv
beautifulsoup4
lxml
scikit-learn
xgboost
streamlit
setuptools==68.0.0
finnhub-python
backtesting
bokeh